payment-method-id/payment-request-ctor-pmi-handling.https.sub.html

<!DOCTYPE html>
<!-- Copyright Â© 2017 Mozilla and World Wide Web Consortium, (Massachusetts Institute of Technology, ERCIM, Keio University, Beihang). -->
<meta charset="utf-8">
<title>Test for validity of payment method identifiers during construction</title>
<link rel="help" href="https://w3c.github.io/browser-payment-api/#constructor">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
"use strict";
const validAmount = Object.freeze({
  currency: "USD",
  value: "1.0",
});
const validTotal = Object.freeze({
  label: "Default Total",
  amount: validAmount,
});
const defaultDetails = Object.freeze({
  total: validTotal,
});

test(() => {
  const validMethods = [
    "https://wpt",
    "https://{{domains[nonexistent]}}/",
    "https://{{domains[nonexistent]}}/payment",
    "https://{{domains[nonexistent]}}/payment-request",
    "https://{{domains[nonexistent]}}/payment-request?",
    "https://{{domains[nonexistent]}}/payment-request?this=is",
    "https://{{domains[nonexistent]}}/payment-request?this=is&totally",
    "https://{{domains[nonexistent]}}:443/payment-request?this=is&totally",
    "https://{{domains[nonexistent]}}:443/payment-request?this=is&totally#fine",
    "https://:@{{domains[nonexistent]}}:443/payment-request?this=is&totally#ðŸ‘",
    " \thttps://wpt\n ",
    "https://xn--c1yn36f",
    "https://é»žçœ‹",
  ];
  for (const validMethod of validMethods) {
    try {
      const methods = [{ supportedMethods: validMethod }];
      new PaymentRequest(methods, defaultDetails);
    } catch (err) {
      assert_unreached(
        `Unexpected exception with valid standardized PMI: ${validMethod}. ${err}`
      );
    }
  }
}, "Must support valid standard URL PMIs");

test(() => {
  const invalidMethods = [
    "https://username@example.com/pay",
    "https://:password@example.com/pay",
    "https://username:password@example.com/pay",
    "http://username:password@example.com/pay",
    "http://foo.com:100000000/pay",
    "not-https://{{domains[nonexistent]}}/payment-request",
    "../realitive/url",
    "/absolute/../path?",
    "https://",
  ];
  for (const invalidMethod of invalidMethods) {
    assert_throws_js(
      RangeError,
      () => {
        const methods = [{ supportedMethods: invalidMethod }];
        new PaymentRequest(methods, defaultDetails);
      },
      `expected RangeError processing invalid URL PMI "${invalidMethod}"`
    );
  }
}, "Constructor MUST throw if given an invalid URL-based payment method identifier");

</script>
